+++ CHANGES.md
@@ -3,6 +3,7 @@
 ### 2.3.2
  * Bugfix: Fix AUCTION symbol parsing on Coinbase
  * Bugfix: Fix PERPETUAL symbol parsing on Phemex
+ * Feature: Added support for Bybit spot orderbook and trade websocket endpoints
 
 ### 2.3.1 (2022-10-31)
  * Bugfix: timestamp not reset correctly on reconnect