+++ CHANGES.md
@@ -3,6 +3,8 @@
 ### 2.3.2
  * Bugfix: Fix AUCTION symbol parsing on Coinbase
  * Bugfix: Fix PERPETUAL symbol parsing on Phemex
+ * Feature: Access to all AIOKafka configuration options
+ * Feature: Use backend Queue for Kafka
 
 ### 2.3.1 (2022-10-31)
  * Bugfix: timestamp not reset correctly on reconnect